Add Week2 Excercise 5: Popups, Loading, Spinner and Gestures

+++ Week2/conFusion/www/js/controllers.js
@@ -190,7 +190,7 @@ angular.module('conFusion.controllers', [])
             
                     }])
 
-        .controller('FavoritesController', ['$scope', 'menuFactory', 'favoriteFactory', 'baseURL', '$ionicListDelegate', function ($scope, menuFactory, favoriteFactory, baseURL, $ionicListDelegate) {
+        .controller('FavoritesController', ['$scope', 'menuFactory', 'favoriteFactory', 'baseURL', '$ionicListDelegate', '$ionicPopup', '$ionicLoading', '$timeout', function ($scope, menuFactory, favoriteFactory, baseURL, $ionicListDelegate, $ionicPopup, $ionicLoading, $timeout) {
 
             $scope.baseURL = baseURL;
             $scope.shouldShowDelete = false;
@@ -200,23 +200,43 @@ angular.module('conFusion.controllers', [])
             $scope.dishes = menuFactory.getDishes().query(
                 function (response) {
                     $scope.dishes = response;
+                    $timeout(function () {
+                        $ionicLoading.hide();
+                        }, 1000);
                 },
                 function (response) {
                     $scope.message = "Error: " + response.status + " " + response.statusText;
+                    $timeout(function () {
+                        $ionicLoading.hide();
+                    }, 1000);
                 });
+            
             console.log($scope.dishes, $scope.favorites);
 
             $scope.toggleDelete = function () {
             $scope.shouldShowDelete = !$scope.shouldShowDelete;
             console.log($scope.shouldShowDelete);
-            }
+            };
 
             $scope.deleteFavorite = function (index) {
-        
-            favoriteFactory.deleteFromFavorites(index);
-            $scope.shouldShowDelete = false;
 
-        }}])
+                var confirmPopup = $ionicPopup.confirm({
+                    title: 'Confirm Delete',
+                    template: 'Are you sure you want to delete this item?'
+                });
+
+                confirmPopup.then(function (res) {
+                    if (res) {
+                    console.log('Ok to delete');
+                    favoriteFactory.deleteFromFavorites(index);
+                    } else {
+                    console.log('Canceled delete');
+                    }
+                });
+
+                $scope.shouldShowDelete = false;
+
+            }}])
 
 .filter('favoriteFilter', function () {
     return function (dishes, favorites) {
